requirements.txt
streamlit==1.37.0
python-pptx==0.6.23
Pillow==10.4.0
openai>=1.40.0
anthropic>=0.34.0
google-generativeai>=0.7.0
